+++ spec/javascripts/api_spec.js
@@ -288,7 +288,7 @@ describe('Api', () => {
     it('creates a group label', done => {
       const namespace = 'group/subgroup';
       const labelData = { some: 'data' };
-      const expectedUrl = `${dummyUrlRoot}/groups/${namespace}/-/labels`;
+      const expectedUrl = Api.buildUrl(Api.groupLabelsPath).replace(':namespace_path', namespace);
       const expectedData = {
         label: labelData,
       };